requirements.txt
pandas==1.1.5
pyarrow==6.0.1
faiss-gpu==1.7.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
pillow==8.4.0
webdataset==0.2.35
numpy==1.19.5
tensorboard
click